+++ lib/delete_recursively.rb
@@ -22,26 +22,59 @@ module DeleteRecursively
   module DependencyHandling
     def handle_dependency
       return super unless DeleteRecursively.enabled_for?(self)
+
       if reflection.belongs_to?
-        # can't use ::dependent_ids, owner is already destroyed at this point
-        ids = load_target ? target.id : []
+        # Special case. The owner is already destroyed at this point,
+        # so we cannot use the efficient ::dependent_ids lookup. Note that this
+        # only happens for a single entry-record on #destroy, though.
+        return unless target = load_target
+
+        DeleteRecursively.delete_records_recursively(target.class, target.id)
       else
-        ids = DeleteRecursively.dependent_ids(owner.class, owner.id, reflection)
+        DeleteRecursively.delete_recursively(reflection, owner.class, owner.id)
       end
-      DeleteRecursively.delete_recursively(reflection, ids)
     end
   end
 
   class << self
-    def delete_recursively(reflection, record_ids)
-      record_class = reflection.klass
-      if recurse_on?(reflection)
-        record_class.reflect_on_all_associations.each do |sub_reflection|
-          sub_ref_ids = dependent_ids(record_class, record_ids, sub_reflection)
-          delete_recursively(sub_reflection, sub_ref_ids)
+    def delete_recursively(reflection, owner_class, owner_ids, seen = [])
+      # Dependent deletion can be bi-directional, so we need to avoid a loop.
+      return if seen.include?(reflection)
+
+      seen << reflection
+
+      associated_classes(reflection).each do |record_class|
+        record_ids =
+          dependent_ids(owner_class, owner_ids, reflection, record_class)
+        if recurse_on?(reflection)
+          record_class.reflect_on_all_associations.each do |subref|
+            delete_recursively(subref, record_class, record_ids, seen)
+          end
         end
+        destroy_or_delete(reflection, record_class, record_ids)
       end
-      destroy_or_delete(reflection, record_class, record_ids)
+    end
+
+    def associated_classes(reflection)
+      if reflection.polymorphic?
+        # This ignores relatives where the inverse relation is not defined.
+        # The alternative would be to expensively select all distinct values
+        # from the *_type column:
+        # reflection.active_record.distinct.pluck(reflection.foreign_type)
+        ActiveRecord::Base.descendants.select do |klass|
+          klass.reflect_on_all_associations
+               .any? { |ref| ref.inverse_of == reflection }
+        end
+      else
+        [reflection.klass]
+      end
+    end
+
+    def delete_records_recursively(record_class, record_ids)
+      record_class.reflect_on_all_associations.each do |ref|
+        delete_recursively(ref, record_class, record_ids)
+      end
+      record_class.delete(record_ids)
     end
 
     def destroy_or_delete(reflection, record_class, record_ids)
@@ -68,13 +101,16 @@ module DeleteRecursively
       %i[delete delete_all].include?(reflection.options[:dependent])
     end
 
-    def dependent_ids(owner_class, owner_ids, reflection)
+    def dependent_ids(owner_class, owner_ids, reflection, assoc_class = nil)
       if reflection.belongs_to?
-        owners_arel = owner_class.where(owner_class.primary_key => owner_ids)
-        owners_arel.pluck(reflection.association_foreign_key).compact
+        owners = owner_class.where(owner_class.primary_key => owner_ids)
+        if reflection.polymorphic?
+          owners = owners.where(reflection.foreign_type => assoc_class.to_s)
+        end
+        owners.pluck(reflection.association_foreign_key).compact
       elsif reflection.through_reflection
         dependent_ids_with_through_option(owner_class, owner_ids, reflection)
-      else
+      else # plain `has_many` or `has_one`
         owner_foreign_key = foreign_key(owner_class, reflection)
         reflection.klass.where(owner_foreign_key => owner_ids).ids
       end
@@ -105,12 +141,18 @@ module DeleteRecursively
       reflection && reflection.foreign_key || owner_class.to_s.foreign_key
     end
 
-    def all(record_class, criteria = {})
+    def all(record_class, criteria = {}, seen = [])
+      return if seen.include?(record_class)
+
+      seen << record_class
+
       record_class.reflect_on_all_associations.each do |reflection|
-        if recurse_on?(reflection)
-          all(reflection.klass, criteria)
-        elsif deleting?(reflection)
-          delete_with_applicable_criteria(reflection.klass, criteria)
+        associated_classes(reflection).each do |assoc_class|
+          if recurse_on?(reflection)
+            all(assoc_class, criteria, seen)
+          elsif deleting?(reflection)
+            delete_with_applicable_criteria(assoc_class, criteria)
+          end
         end
       end
       delete_with_applicable_criteria(record_class, criteria)
@@ -127,10 +169,26 @@ end
 
 require 'active_record'
 
-%w[BelongsTo HasMany HasOne].each do |assoc_name|
-  assoc_builder = ActiveRecord::Associations::Builder.const_get(assoc_name)
-  assoc_builder.singleton_class.prepend(DeleteRecursively::OptionPermission)
+module ActiveRecord
+  module Associations
+    %w[BelongsTo HasMany HasOne].each do |assoc_name|
+      assoc_builder = Builder.const_get(assoc_name)
+      assoc_builder.singleton_class.prepend(DeleteRecursively::OptionPermission)
 
-  assoc_class = ActiveRecord::Associations.const_get("#{assoc_name}Association")
-  assoc_class.prepend(DeleteRecursively::DependencyHandling)
+      assoc_class = const_get("#{assoc_name}Association")
+      assoc_class.prepend(DeleteRecursively::DependencyHandling)
+    end
+  end
+
+  class Base
+    def delete_recursively
+      DeleteRecursively.delete_records_recursively(self.class, id)
+    end
+  end
+
+  class Relation
+    def delete_all_recursively
+      DeleteRecursively.delete_records_recursively(klass, ids)
+    end
+  end
 end