+++ spec/dummy_rails_app/app/models/blog.rb
@@ -1,3 +1,4 @@
 class Blog < ApplicationRecord
   has_many :posts, dependent: :delete_recursively
+  before_destroy { logger.debug 'Blog destroy callback!' }
 end