+++ spec/dummy_rails_app/app/models/delivery_service.rb
@@ -1,3 +1,4 @@
 class DeliveryService < ApplicationRecord
   has_many :pizzas, dependent: :delete_recursively
+  has_many :prices, dependent: :delete_recursively
 end