+++ spec/dummy_rails_app/app/models/doomsday_device.rb
@@ -1,0 +1,3 @@
+class DoomsdayDevice < ApplicationRecord
+  has_one :price, as: :product, dependent: :delete_recursively
+end