+++ spec/dummy_rails_app/app/models/price.rb
@@ -1,0 +1,4 @@
+class Price < ApplicationRecord
+  belongs_to :delivery_service
+  belongs_to :product, polymorphic: true, dependent: :delete_recursively
+end